Cycle: m0-truth-first-foundation
- 5 spec-mandated commits (reconstruction corpus, ROADMAP, coverage CI, m0_acceptance stubs, LEGACY disposition)
- 6 extension commits (pre-existing clippy/fmt cascade, AGENTS.md, gitignore fix)
- Vault artifacts: 13 ADRs, 7 specs, 10 m0-NN tickets
- Verify: PASS (6/6 lenses)
- Debt-verify: PASS (5/5 clusters)

Refs: vault/cycles/m0-truth-first-foundation/{proposal,spec,design,tasks,apply-checkpoint,verify,debt-verify}

+++ crates/chronos-domain/src/index/shadow.rs
@@ -106,7 +106,7 @@ mod tests {
     fn test_many_addresses() {
         let mut index = ShadowIndex::new();
         for i in 0..1000 {
-            index.insert(i as u64 * 8, i);
+            index.insert(i * 8, i);
         }
         assert_eq!(index.unique_addresses(), 1000);
         assert_eq!(index.len(), 1000);
